Use long instead of int

+++ cli/src/main/java/keywhiz/cli/commands/AssignAction.java
@@ -88,13 +88,13 @@ public class AssignAction implements Runnable {
         }
 
         try {
-          if (keywhizClient.groupDetailsForId(Math.toIntExact(group.getId())).getClients().contains(client)) {
+          if (keywhizClient.groupDetailsForId(group.getId()).getClients().contains(client)) {
             throw new AssertionError(
                 format("Client '%s' already assigned to group '%s'", assignActionConfig.name,
                     group.getName()));
           }
           logger.info("Enrolling client '{}' in group '{}'.", client.getName(), group.getName());
-          keywhizClient.enrollClientInGroupByIds(Math.toIntExact(client.getId()), Math.toIntExact(group.getId()));
+          keywhizClient.enrollClientInGroupByIds(client.getId(), group.getId());
         } catch (IOException e) {
           throw Throwables.propagate(e);
         }
@@ -103,7 +103,7 @@ public class AssignAction implements Runnable {
       case "secret":
 
         try {
-          int groupId = Math.toIntExact(group.getId());
+          long groupId = group.getId();
           String[] parts = splitNameAndVersion(assignActionConfig.name);
           SanitizedSecret sanitizedSecret =
                 keywhizClient.getSanitizedSecretByNameAndVersion(parts[0], parts[1]);
@@ -113,7 +113,7 @@ public class AssignAction implements Runnable {
                     group.getName()));
           }
           logger.info("Allowing group '{}' access to secret '{}'.", group.getName(), sanitizedSecret.name());
-          keywhizClient.grantSecretToGroupByIds(Math.toIntExact(sanitizedSecret.id()), groupId);
+          keywhizClient.grantSecretToGroupByIds(sanitizedSecret.id(), groupId);
         } catch (KeywhizClient.NotFoundException e) {
           throw new AssertionError("Secret doesn't exist.");
         } catch (ParseException e) {
